+++ content/adapters/liepin.js
@@ -61,7 +61,7 @@ class LiepinAdapter extends BaseAdapter {
     };
   }
 
-  parseSearchResults() { return this.getJobElements().map(el => this.extractJobInfo(el)); }
+  parseSearchResults() { let els = this.getJobElements(); if (!els.length) els = this._getAllPossibleCards(document); return els.map(el => this.extractJobInfo(el)); }
 
   async applyToPosition(element) {
     try {